# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Employee Data Access API 2
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

paths:
  /postToken:
    post:
      description: Post token for authentication
      x-swagger-router-controller: employees
      operationId: postToken

      parameters:

        - in: body
          name: body
          description: Add a user
          required: false
          schema:
            properties:
              token:
                type: string
            
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /employees:
    get:
      description: Get Employees
      x-swagger-router-controller: employees
      operationId: getEmployees
      
      parameters:
        -
          name: size
          in: query
          description: Size of array
          required: true
          type: number
          format: double
        -
          name: user
          in: query
          description: Return Employees belonging to user
          required: false
          type: string
        - 
          name: token
          in: header
          description: Authorisation Token
          required: true
          type: string  

      responses:
        "200":
          description: Successful response
          # Use JSON Schema format
          schema:
            $ref: "#/definitions/EmployeeSchema"

    post:
      description: Post Employee (INSERT)
      x-swagger-router-controller: employees
      operationId: createEmployee

      parameters:
        - in: body
          name: body
          description: Add a Employee
          required: true
          schema:
            $ref: "#/definitions/EmployeeSchema"
            
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /employees/{id}: 
    get:             
       description: Returns a employee based on a passed id
       x-swagger-router-controller: employees
       operationId: getSingleEmployee
       
       parameters: 
         -
           name: id
           in:  path
           description: ID of employee to fetch
           required: true
           type: string

       responses:
         "201":
            description: Null response
         default:
           description: unexpected error
           schema:
             $ref: "#/definitions/ErrorResponse"            

    put:
      description: Put Employee (UPDATE)
      x-swagger-router-controller: employees
      operationId: updateEmployee
       
      parameters:
        - 
          name: id
          in:  path
          description: Update a Employee
          required: true
          type: string
          
        - name: body
          in: body
          description: Employee to update
          required: false
          schema:
            $ref: "#/definitions/EmployeeSchema"
  
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Employee not found
        "405":
          description: Validation exception   
          
    delete: 
      description: deletes a single employee
      x-swagger-router-controller: employees
      operationId: deleteEmployee

      parameters: 
        - 
          name: id
          in: path
          description: ID of employee to delete
          required: true
          type: integer
          format: int64
      
      responses: 
        "204":
          description: employee deleted
        default: 
          description: unexpected error
          schema: 
            $ref: "#/definitions/ErrorResponse"         

  /swagger:
    x-swagger-pipe: swagger_raw
    
# complex objects have schema definitions
definitions:

  EmployeeSchema:
  
    required:
    
      - userId
      - fName
      - lName
      - role
      
    properties:
    
      userId:
        type: string
      fName:
        type: string        
      lName:
        type: string
      addr1: 
        type: string
      addr2:
        type: string
      role:
        type: string

  ErrorResponse:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string  